requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
neo4j>=5.14.0
openai>=1.3.0
pyyaml>=6.0.1
pydantic>=2.5.0
python-dotenv>=1.0.0
